it works but much cleanup is needed
