+++ rust/headless-client/Cargo.toml
@@ -16,6 +16,7 @@ firezone-cli-utils = { workspace = true }
 futures = "0.3.30"
 git-version = "0.3.9"
 humantime = "2.1"
+ip_network = { version = "0.4", default-features = false }
 secrecy = { workspace = true }
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
@@ -29,8 +30,11 @@ url = { version = "2.3.1", default-features = false }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 dirs = "5.0.1"
+netlink-packet-route = { version = "0.19", default-features = false }
+netlink-packet-core = { version = "0.7", default-features = false }
 nix = { version =  "0.28.0", features = ["fs", "user"] }
 resolv-conf = "0.7.0"
+rtnetlink = { workspace = true }
 sd-notify = "0.4.1" # This is a pure Rust re-implementation, so it isn't vulnerable to CVE-2024-3094
 
 [target.'cfg(target_os = "macos")'.dependencies]
